Fixing compilation errors (#2)

+++ plugins/sqltools/org.eclipse.datatools.sqltools.result/src/org/eclipse/datatools/sqltools/result/internal/index/ResultHistoryLuceneIndex.java
@@ -221,7 +221,7 @@ public class ResultHistoryLuceneIndex implements IResultHistoryIndex
                 Query query = parser.parse(expression);
                 IndexSearcher searcher = new IndexSearcher(reader);
                 TopDocs hits = searcher.search(query, reader.maxDoc());
-                int count = hits.totalHits;
+                int count = (int)hits.totalHits;
                 IResultInstance[] instances = new IResultInstance[count];
                 
                 for(int i=0;i<count;i++)
